refactor 一旦全部切り出し
